1:1 match pbrt-v4 reference across crown / killeroo / materials / bunny_cloud
The Crown scene rendered horizontally mirrored because
parse_pbrt_string's LookAt handler REPLACED the CTM with `pbrt_lookat(...)`
output, dropping any prior transform. pbrt-v4 BasicSceneBuilder::LookAt
multiplies onto the CTM, so Crown's leading `Scale -1 1 1` propagated into
the camera transform. Fix:

    ctm = ctm * pbrt_lookat(eye, target, up)

BumpMap was using normalized shading tangent + missing the `h * dndu` term,
so curved bumped surfaces (Crown gold dome) rendered nearly flat.
perturb_bump_frame now follows pbrt-v4 materials.h:BumpMap exactly:

    dpdu' = dpdu + (∂h/∂u) * n + h * dndu
    dpdv' = dpdv + (∂h/∂v) * n + h * dndv

with unnormalized dpdu/dpdv (shapes.h:959 `ss = isect.dpdu`) and FaceForward
against the GEOMETRIC normal (interaction.cpp:184-187 `SetShadingGeometry`).
The wrapper around sample_bsdf/evaluate_bsdf is removed; the perturbation
is hoisted to intersection time via get_perturbed_shading_frame so the
integrator's cos factors and direct-lighting MIS see the bumped frame.

BUMP_DEFAULT_DELTA dropped 1/256 → 5e-4 to match pbrt-v4 default; the
larger value was averaging out high-frequency ornamental engraving.

Compute per-triangle dndu/dndv via Cramer's rule on vertex normals so the
`h * dndu` term works on curved meshes.

Shadow ray cos_theta now uses shading normal `ns`; ray offset still uses
geometric `ng` (pbrt-v4 OffsetRayOrigin convention). Without this, bumped
surfaces lost their angular falloff and arealight contribution stayed
constant across grazing angles.

scene_builder:
- aspect-aware screen window so non-square films don't get squashed
- parse `lensradius` / `focaldistance` for depth-of-field
- decode 8-bit imagemaps as sRGB (pbrt-v4 default)
- mix material `amount` as texture, coated-diffuse `roughness` as texture
- pixelsamples from the Sampler block

Surface eval scale now uses PERSPECTIVE_NEAR (was `sqrt(spp)`); bilinear
texture filtering (was nearest); volpath intersection threads real ray
differentials through bump.

Tests: new shadow_ test scenes + pbrt EXR references; "Cast shadows"
testset added to test_pbrt_all_materials.jl. All 4 benchmark scenes
match pbrt orientation; Crown SAD vs pbrt at 512 spp ≈ 0.101.

+++ src/materials/coated-diffuse.jl
@@ -103,11 +103,11 @@ function CoatedDiffuse(;
     n_samples::Int = 1,
     remap_roughness::Bool = true
 )
-    # Handle roughness - can be scalar or (u,v) tuple
+    # Handle roughness - can be scalar, (u,v) tuple, or Texture{Float32}
     u_rough, v_rough = if roughness isa Tuple
-        Float32(roughness[1]), Float32(roughness[2])
+        roughness[1], roughness[2]
     else
-        Float32(roughness), Float32(roughness)
+        roughness, roughness
     end
 
     CoatedDiffuse(
